+++ src/components/ui/PopOver/index.tsx
@@ -6,13 +6,15 @@ import React, {
   createRef,
   useState,
   HTMLAttributes,
-  useEffect
+  useEffect,
 } from 'react';
 import { Manager, Reference, Popper } from 'react-popper';
+import classnames from 'classnames';
 
 import { KEY_CODES } from '../../../constants';
 import useClickOutside from '../../../hooks/useClickOutside';
 import useTabOutside from '../../../hooks/useTabOutside';
+import { BaseProps } from '../Base';
 import { StyledPopOverMenu, StyledPopOverToggle } from './Styled';
 
 export type Placement =
@@ -25,17 +27,22 @@ export type Placement =
   | 'left-start'
   | 'left-end';
 
-export interface PopOverProps extends HTMLAttributes<HTMLSpanElement> {
+export interface PopOverProps
+  extends Omit<HTMLAttributes<HTMLUListElement>, 'css'>,
+    BaseProps {
   /** CSS classname to apply custom styles. */
   className?: string;
   /** Whether or not this is a sub menu. */
   isSubMenu?: Boolean;
   /** Defines the placement of PopOver menu. */
   placement?: Placement;
-  /** Defines the function to render button(s). */
-  renderButton: (isActive: boolean) => {};
+  /** Defines the function to render the inner contents of the popover button element */
+  renderButton?: (isActive: boolean) => {};
+  /** Alternative to renderButton, defines the function to render the full popover button element (as opposed to just its contents). This is used if you want full control over the button rendering. The button must forwardRef */
+  renderButtonWrapper?: (isActive: boolean, props: any) => {};
   /** The label used for availability. */
   a11yLabel: string;
+  /** The elements that populate the menu */
   children: any;
 }
 
@@ -45,11 +52,13 @@ const getFocusableElements = (node: HTMLElement): NodeListOf<HTMLElement> => {
 
 export const PopOver: FC<PopOverProps> = ({
   renderButton,
+  renderButtonWrapper,
   children,
   isSubMenu = false,
   placement = 'bottom-start',
   a11yLabel,
   className,
+  ...rest
 }) => {
   const menuRef = createRef<HTMLSpanElement>();
   const [isOpen, setIsOpen] = useState(false);
@@ -111,25 +120,38 @@ export const PopOver: FC<PopOverProps> = ({
     <span ref={menuRef} onKeyDown={handleKeyUp} data-testid="popover">
       <Manager>
         <Reference>
-          {({ ref }) => (
-            <StyledPopOverToggle
-              data-menu={isSubMenu ? 'submenu' : null}
-              onClick={() => setIsOpen(!isOpen)}
-              ref={ref}
-              aria-label={a11yLabel}
-              aria-haspopup={true}
-              aria-expanded={isOpen}
-              data-testid="popover-toggle"
-              className={className || ''}
-            >
-              {renderButton(isOpen)}
-            </StyledPopOverToggle>
-          )}
+          {({ ref }) => {
+            const props = {
+              ref,
+              className: classnames(className, 'ch-popover-toggle'),
+              onClick: () => setIsOpen(!isOpen),
+              'data-menu': isSubMenu ? 'submenu' : null,
+              'aria-label': a11yLabel,
+              'aria-haspopup': true,
+              'aria-expanded': isOpen,
+              'data-testid': 'popover-toggle',
+            };
+
+            if (renderButton) {
+              return (
+                <StyledPopOverToggle {...props}>
+                  {renderButton(isOpen)}
+                </StyledPopOverToggle>
+              );
+            }
+
+            if (renderButtonWrapper) {
+              return renderButtonWrapper(isOpen, props);
+            }
+
+            return null;
+          }}
         </Reference>
         {isOpen && (
           <Popper
             placement={placement}
             modifiers={[{ name: 'offset', options: { offset: [-8, 0] } }]}
+            {...rest}
           >
             {({ ref, style, placement }) => (
               <StyledPopOverMenu