+++ src/components/ui/icons/All/All.stories.tsx
@@ -8,6 +8,7 @@ import { text } from '@storybook/addon-knobs';
 import Add from '../Add';
 import Arrow from '../Arrow';
 import Attendees from '../Attendees';
+import Attachment from '../Attachment';
 import Camera from '../Camera';
 import Caret from '../Caret';
 import Caution from '../Caution';
@@ -17,18 +18,24 @@ import CheckRound from '../CheckRound';
 import Clear from '../Clear';
 import Clock from '../Clock';
 import Cog from '../Cog';
+import Dislike from '../Dislike';
+import Dock from '../Dock';
 import Dots from '../Dots';
 import ScreenShare from '../ScreenShare';
 import Crown from '../Crown';
 import DeskPhone from '../DeskPhone';
 import Dialer from '../Dialer';
+import Document from '../Document';
 import Hamburger from '../Hamburger';
 import Information from '../Information';
 import Laptop from '../Laptop';
 import LeaveMeeting from '../LeaveMeeting';
+import Like from '../Like';
+import Lock from '../Lock';
 import Meeting from '../Meeting';
 import Microphone from '../Microphone';
 import Pause from '../Pause';
+import Pin from '../Pin';
 import Phone from '../Phone';
 import Play from '../Play';
 import Presenter from '../Presenter';
@@ -40,6 +47,10 @@ import Share from '../Share';
 import SignalStrength from '../SignalStrength';
 import Sound from '../Sound';
 import Flex from '../../Flex';
+import Eye from '../Eye';
+import EmojiPicker from '../EmojiPicker';
+import UpAndDownCaret from '../UpAndDownCaret';
+import Spinner from '../Spinner';
 
 export default {
   title: 'UI Components/Icons/All',
@@ -50,6 +61,7 @@ const icons = [
   Add,
   Arrow,
   Attendees,
+  Attachment,
   Camera,
   Caret,
   Caution,
@@ -59,18 +71,27 @@ const icons = [
   Clear,
   Clock,
   Cog,
+  Crown,
+  Dislike,
   Dots,
+  Dock,
+  Document,
   ScreenShare,
   Crown,
   DeskPhone,
   Dialer,
+  EmojiPicker,
+  Eye,
   Hamburger,
   Information,
   Laptop,
   LeaveMeeting,
+  Like,
+  Lock,
   Meeting,
   Microphone,
   Pause,
+  Pin,
   Phone,
   Play,
   Presenter,
@@ -81,6 +102,8 @@ const icons = [
   Share,
   SignalStrength,
   Sound,
+  Spinner,
+  UpAndDownCaret
 ];
 
 const IconWrapper = styled.div`